rename `FileContext` to `FileProvider`
