+++ src/ui/menus/BottomBar/SyncState.tsx
@@ -1,7 +1,7 @@
 import { ErrorOutline } from '@mui/icons-material';
 import { CircularProgress, Tooltip, useTheme } from '@mui/material';
 import { ShowAfter } from '../../../components/ShowAfter';
-import { useFileContext } from '../../contexts/FileContext';
+import { useFileContext } from '../../components/FileProvider';
 import BottomBarItem from './BottomBarItem';
 
 export default function SyncState() {
